Version 2.0
This version is a major improvement over the first. It now has a working
microkernel that can now recieve user input. It also has a branch
predictor, caches, and many more realistic features added. These are all
features I learned in my computer architecture class.

There is still a lot more that can be added but I am extremely happy
with the progress I have made for this project. I plan to take a more
advanced computer architecture class in the winter semester so I will
hopefully be able to continue to add more complex features to this
simulation.